+++ save_game.py
@@ -9,8 +9,10 @@ Date: 10/26/2024
 Last Modified: 10/26/2024
 '''
 
+# saves the current game to the save file (currently hardcoded to save.txt)
 def save(ui_manager):
     with open('save.txt', 'w') as file:
+        # saves the user's current balance along with any shop purchases they have made
         output = f'{ui_manager.cookie_count}\n'
         for item in ui_manager.shop_items.values():
             output += f'"{item.name}":{item.purchased_count}\n'