+++ Documentation/ChangeLog.md
@@ -4,11 +4,13 @@
 ### Added
 
 - load_game code to populate upgrades_acquired list so that the upgrades are shown after loading a save **(Jack)**
+- Documentation for all of the game's files and functions **(Peter)**
 
 ### Modified
 
 - handle_shop_click to check if an item is already in upgrades acquired to avoid filling the UI with duplicate upgrades **(Jack)**
 - draw_upgrades to display cpc or cps depending on the item **(Jack)**
+
 ## October 26 (Sprint 1)
 
 ### Added